Changed operations to respect verbosity options.

+++ delete.py
@@ -1,6 +1,7 @@
 import os
 
-def Delete (names):
+def Delete (names, ctlFunc = lambda s, d: True):
     """ Takes in list of files. Deletes each file. """
     for n in names:
-        os.remove(n) 
+        if ctlFunc(n, "*DELETE*"):
+            os.remove(n)
